Use the modern page model/view settings for all pages.
Review Board 4.0 introduced a new set of template blocks for setting the
JavaScript page model/view and any attributes or options needed by the
page. This wasn't being used by some of the pages, resulting in a
warning in the debug console.

This change switches over the review request page, diff viewer, and
review UIs to use the modern blocks.

Most of this is pretty mechanical. One change you will notice is that we
are no longer manually injecting the `lastActivityTimestamp` with a
`{% localtime off %}` surrounding it. This hasn't actually been needed
in a very long time. Using `{% localtime off %}` will impact template
filters like `|date`, which we used to use once upon a time. We've since
moved to using the JSON serializer, which always encodes as UTC, so this
isn't needed anymore. Because of this, this value has simply been moved
into the common page data results that are fed into the model.

Testing Done:
Manually tested the review request page, diff viewer, and the image
review UI, making sure they were all working and that the correct state
was getting passed in to the models.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10916/