Fix a regression from the new RB.UploadAttachmentView change.
The change to `RB.UploadAttachmentView` introduced in
e94bac071aa42fa45ad30d92f975d6d7c114e16a was landed prematurely, and had
a couple of regressions.

The renamed `fileAttachmentThumbnailView.es6.js` was missing in
`staticbundles.py`, and a few calls to `render()` should have been
updated to `show()`.

This is a trivial change that fixes these issues.