Fix build-i18n.py to set up the appropriate environment before importing.
`build-i18n.py`, which is a necessary part of the package build process,
was assuming it had the right version of Review Board and Django in the
import path when it began importing modules. While this is often going
to be the case when building packages, it's not necessarily the case in
CI.

Our `build-media.py` script has code to set up the correct environment
before doing imports, so it was able to work fine in CI. This change
simply copies over that logic to `build-i18n.py`.

Testing Done:
Built the package successfully.

Reviewed at https://reviews.reviewboard.org/r/10944/