Add a new RB.Product constant for product information.
This introduces `RB.Product`, which contains information on the product
name, version (human-readable and raw version information), release
status, and manual URL. This can be used by the JavaScript codebase for
some purposes (like looking up the manual URL, which was previously
doable with an old constant that's now deprecated), but its main use is
for extensions.

Extensions that need to alter behavior client-side based on the version
of Review Board can now do so by reading `RB.Product.versionInfo`. This
can be important given the differences between Review Board 3.0 and 4.0,
particularly in the administration UI.

The information is generated once per process and then cached in-process
for future lookups.

Testing Done:
Loaded a couple pages and verified that `RB.Product` was generated and
contained the correct information.

Reviewed at https://reviews.reviewboard.org/r/10947/