Update the notification for browsers with JavaScript turned off.
When browsers with JavaScript disabled try to access Review Board, we
show a warning about the lack of JavaScript, and offer some guidance.
This code is very old, and we single out a solution for Firefox without
considering the modern world of JavaScript in browsers and other
possible extensions.

This change updates this part of the page, using a modern `rb-c-alert`
and simplifying the information shown. We now show this in the navbar
area, rather than placing it in the main content area (which may be
set to hide until JavaScript shows it, for certain pages). This ensures
it's always visible, and that it's docked toward the top of every page.

Testing Done:
Turned off JavaScript in the developer tools and visited a couple of
pages. Saw the new banner.

Reviewed at https://reviews.reviewboard.org/r/10948/