Define a global-vars block for all global variable definitions.
This is a simple change that introduces a new `global-vars` block in the
base template, which child templates can use to register global
variables (through the new `{% definevar "..." global %}` support in
Djblets). Any global variables defined here will be usable in any other
block in the template, and are a good place to define IDs, URLs, or
blocks of HTML that need to be used in multiple independent blocks.

The new block will be used in an upcoming change shortly.

Testing Done:
Made use of this in an upcoming change. Verified that the variables were
usable in other blocks.

Reviewed at https://reviews.reviewboard.org/r/10954/