Serialize and allow setting extra_data for repositories in the API.
The repositories API didn't allow any access to the `extra_data` field,
either reading from the field or setting values in the field. This made
it difficult to look up important settings that might affect a client,
or to set any custom values.

This change exposes the field to the caller, allowing them to see all
the settings (custom and tied to a SCMTool/hosting service), and to
modify it using standard means (directly setting fields or using JSON
Patches or Merge Patches).

Testing Done:
All unit tests pass.

Tested by a customer who was building an integration that needed this.
Verified on their end that the fields in `extra_data` could be safely
manipulated.

Reviewed at https://reviews.reviewboard.org/r/10961/