Fix typos in fields.py comments.
In `reviewboard/reviews/fields.py`, there are two typos in the comments.
One is in the comments for method `propagate_data()`: `explictly` should
be `explicitly`. And the other one is `appropraite` (should be
`appropriate`) in the comments for `get_css_classes()`.

Testing Done:
N/A

Reviewed at https://reviews.reviewboard.org/r/10927/