Update the model deletion confirmation page in the admin UI.
This adds a custom template for the delete confirmation page used in the
admin UI for models. Django's default template doesn't render quite
well, and doesn't offer any flexibility for customization. Our version
is a bit friendlier and fits our modern style in the admin UI.

This covers the main confirmation page, the "protected objects"
confirmation page (used if a `ForeignKey` is set to prevent deletion of
the referenced object using the `PROTECT` flag), or if the user doesn't
have permission for one or more objects.

Testing Done:
Tested the standard deletion page, the permission errors page, and the
protected relations page.

Tested all the buttons on all pages.

Reviewed at https://reviews.reviewboard.org/r/10975/