Add a search field to the repository resource.
The new repository selector UI was using a nonexistant `q` field to
query the repository API. This often appeared to work correctly because
the selector would end up getting the first page of results, and then
selectize would do the filtering on the client side. If the repository
the user wanted wasn't in the first page of results, they wouldn't be
able to get to it.

While the API does have a `name` field, that field tests for equivalence
rather than being a search term like the `q` fields that are part of the
user and group APIs. This change adds a matching `q` field to the
repository API which does an istartswith query on the name.

Testing Done:
- Interacted with the related repository selector widget and checked the
  responses in the network tab.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/10997/