Add utility functions for working with the default SSH storage backend.
This introduces `set_ssh_storage_backend_path()` and
`get_ssh_storage_backend_path()`, which are wrappers around the
`ssh_storage_backend` siteconfig setting. Right now, these don't do a
whole lot, but they give us a standard way for extensions to set a new
backend, and eventually can let us better apply this to the environment
for rbssh (which is not planned for 3.0.x, as it's a more involved
project).

Extensions will also be able to check for the existence of these
functions in order to determine whether to set the old
`settings.RBSSH_STORAGE_BACKEND` or to set the new key in siteconfig.

Testing Done:
Tested these with a patch to Power Pack.

Reviewed at https://reviews.reviewboard.org/r/11008/