Enable use of the new babel-plugin-django-gettext for JavaScript.
This is a simple change to our Babel settings that enables use of our
new `babel-plugin-django-gettext`, used for easier localization. Currently,
no Review Board code makes use of this, but it'll allow us to slowly begin
porting code or using it in new code.

Note that to make use of this, this will require a recent run of
`./setup.py develop`.

Testing Done:
Modified some JavaScript temporarily to make use of the new plugin.
Verified the generated JavaScript was correct.

Reviewed at https://reviews.reviewboard.org/r/11018/