Fix a call to ngettext, and update to use the new plugin capabilities.
During the admin UI rewrite, I forgot a parameter to `ngettext()`. This
previously went unnoticed, but with our new babel-plugin-django-gettext,
it now becomes a build-time error.

This change fixes this up to provide the third argument, the count, and
at the same time switches to using the new syntax, simplifying the code
and reducing the setup for it.

Testing Done:
Verified the generated code (when used against an upcoming build of
babel-plugin-django-gettext, which fixes ngettext-related issues).

Reviewed at https://reviews.reviewboard.org/r/11022/